Stats: Nonstationarity correction fix for non-finite input data
In elements where the GLM cannot be fit to the input data due to too much of it having non-finite values, the enhanced empirical statistic will never exceed zero. Because later during permutation testing the enhanced statistics are dividied by this empirical statistic, this was leading to NaN values in the null distribution. This change sets the empirical statistic to infinity if the empirical statistic could not be derived, so that the subsequent division leads to an enhanced statistic of zero in such elements.