use #pragma GCC diagnostic directives to silence compiler warnings
This is required to silence warnings from within the Eigen or Qt headers
when including with -I rather than -isystem. Only a few places needed
these directives, the rest of the code was modified to simply not bother
including headers that are included in later headers anyway, or changing
the order of inclusion to ensure the main offending headers were already
included (typically within "core/types.h"), and hence would not be
brought it unguarded by a direct inclusion - most of the time, that
simply mean moving Eigen or Qt include directives till after the MRtrix
header includes.