mrconvert -vox: Permit single value input
Prior to this change, providing a single floating-point value as input to the -vox option in mrconvert would result in rescaling of the voxel size along the first axis only. This change makes this behaviour more consistent with the -voxel option in mrresize: If a single value is provided rather than a comma-separated list, set the voxel size to that value along all three spatial axes.