command_history: Only use quotes when needed
When adding the current executing command to the command_history header key-value entry, entries in argv[] are surrounded by quotes in order to deal with file system paths containing special characters. This change attempts to determine when these quotation marks are not necessary and omits them accordingly, to improve the readability of this field.