Major changes to Python scripting library
Previously, executable scripts would call a set of functions within the mrtrix3.app module in the appropriate order in order to initialise the consistent command-line interface, parsing, etc.. A copy-and-paste block of code at the top of each executable file would ensure that this module was present within the Python path.
This change makes the executable interface more similar to that of the C++ files within the cmd/ folder of MRtrix3. Two functions must be defined within the executable file: 'usage()' and 'execute()'. Then, at the bottom of the executable file, the copy-and-paste code block for locating and importing the mrtrix3.app module appears, but also calls the app.execute() function. This function defined within the app module sets up the interface, and uses the Python inspect module to locate and run the two functions defined within the executable file as appropriate.
A major advantage of this approach is that it allows Python exceptions to be thrown by library functions (instead of calling the deprecated app.error() function), which may optionally be caught by the calling code in various circumstances. Standard Python exceptions can also be caught by the app module, and displayed in a manner consistent with the script interface. Additionally, when commands executed via the mrtrix3.run module fail, a dedicated exception class is thrown, which allows both the catching of failed commands in the executing code where permissible, and moves all of the code responsible for handling such command failures to the app module (previously the run module would write a text file, which the app module would then load).
A couple of generic capabilities (e.g. reading the MRtrix3 config file(s)) have been moved to __init__.py, such that they are more easily accessible when using the library but not exploiting the app module interface.
Various changes to the mrtrix3.run module have been made to support multi-threaded use of the run.command() function. This includes use of tempfile.mkstemp() rather than tempfile.TemporaryFile(), as the latter was found to not be thread-safe in Python3 on Windows. The run.command() function additionally now supports list input, where individual entries may themselves be lists (allowing e.g. concatenation of multiple file paths into a command call, but the terminal display will show only the common prefix/suffix of the list rather than the entire contents). run.command() now additionally supports '||' and '&&' operators.