foreach: Port to Python
This commit contains a re-implementation of the 'foreach' script using the MRtrix3 Python scripting library, as opposed to the original implementation which is simply a Bash script. This means that the script will be included in the auto-generated documentation. It also means that the terminal output from multiple running commands can be better controlled.
This does however involve a change to the user interface of this command: Rather than placing a colon character as a 'no-op' to separate the 'foreach' invocation from the command to be executed multiple times, this new version instead expects the final positional command-line argument input to be a string that contains the command to be executed (including any substitution flags). The documentation has been updated to reflect this.