foreach: Make use of subprocess shell=true
In run.command(), accept a parameter that specifies that the command string should be passed to subprocess as-is, and interpreted by a shell spawned specifically for that command, as opposed to the default handling. This ensures that any more advanced usages of the foreach script can be executed as the user intended (following string substitution), provided that any shell operators are appropriately escaped.