run.command(): Deal with lack of write access to /tmp/
In some usage scenarios, the user may not have write privilege to the location that is chosen by default by the tempfile module to create temporary text files (which are used here to store the stdout / stderr outputs of executing commands). This change deals with this prospect by catching the resulting exception and re-running tempfile.mkstemp() while manually specifying the filesystem location in which to create the file, in a location where the user is almost guaranteed to have write access.