New Python class app.runCommandList
This class can be used when there is a series of commands to be executed, the total number of commands to be executed is known, but construction and management of a progressBar would be inconvenient (e.g. the commands are not within an iterative loop). This class manages the construction and update of a progressBar whilst a series of commands are executed, such that the user is presented with a sensible progress message at the terminal rather than a long list of commands where the actual details of the specific commands are not relevant.