Pbrt export for tckconvert
Exports streamlines of length 8 or more into the PBRT Curve format.  See
https://www.pbrt.org/ for details on the rendering system, and
https://www.pbrt.org/fileformat-v3.html#shapes, specifically "curve"
shape.  The PBRT book has more details on curves:
http://www.pbr-book.org/3ed-2018/Shapes/Curves.html

Note: the curves objects have "splitdepth" set to 0 to disable splitting.
This greatly reduces the memory requirements for curves with little
visual change.  The writer class also honors the -increment option
to only export every `increment` points of the streamline.  This option
also reduces the number of curves and reduces memory and computation
time for the PBRT renderer.