New variable MR::App::EXAMPLES
This class allows a command author to insert one or more demonstrative examples of use of the command into the command's help page and online documentation.
It currently only applies to C++ binaries, but an equivalent functionality for Python is expected in the future.
Examples for a limited set of commands that regularly cause user confusion have been added.