DICOM: be more lenient with slice timing vector is Siemens mosaics
There are examples now of Siemens mosaics where the number of entries in
the CSA NumberOfImagesInMosaic entry reports a large size than the
number of slices - leading to the whole entry being ignored - yet the
last few entries are empty. This change allows for this situation, by
taking as many entries as there are slices, and ignoring the last few
entries.