dwibiascorrect: Use algorithm module
In addition to porting to the algorithm module, the output of the fsl algorithm was changed slightly. Instead of directly using the output field from fast, which contains a value of 1.0 (i.e. no correction) outside of the mask, and hence introduces a subtle 'step' in image intensities at the outer edge of this mask, this algorithm will now instead apply a hard mask to the output DWI based on the input mask.