dwipreproc: Compatibility with FSL6.0.0
In this particular version of FSL, the image output by the --fout option of the topup command generates a 4D image, where the first volume is the expected field map image, but the other volumes are intensity-scaled versions of the uncorrected input images. While this is certainly an FSL bug, the dwipreproc script nevertheless needs to be compatible with such data. Currently it causes an issue only when explicit DWI volume recombination is to be performed. This change detects the erroneous topup output, and extracts just the first volume.
Closes #1493.