Streamline::calc_length(): Make no assumptions
Previous version of this function estimated the step size from the central segment of the streamline, and then assumed that this step size applied to the rest of the streamline (apart from the first and last segments) by calling the Streamline::calc_length(float step_size) function. If streamlines have been non-linearly warped, this is not going to be a safe assumption. Therefore, both in any context where a fixed step size is not known, or in tckstats if the user specifies the -explicit option, explicitly sum the distances between streamline vertices.