fixelcfestats: Encapsulate fixel index remapping
When operating using a fixel mask, only fixels from the input template that are within the mask are mapped into the internal data matrix, in such a way that there are no gaps or zero-fills in the data. This requires a mapping from 'external' fixel index (i.e. from the fixel template on file) to 'internal' fixel index (i.e. where data for that fixel lies within the data matrix), and vice-versa.
This functionality has been moved from individual global members in cmd/fixelcfestats.cpp to a class in src/stats/cfe.h, in order to be accessible to other commands.