fixelcfestats: Move handling of non-connected fixels
Rather than writing a self-connection strength of 1.0 into both the normalised connectivity matrix and the smoothing matrix at point of construction, leave them out for the sake of the normalise_matrix() function. Instead, add them explicitly within cmd/fixelcfestats.cpp (since it is the appropriate handling of data in that context where this step is necessary). Also, issue a warning to the user if there are unconnected fixels being processed and either the normalised CFE expression or permutation-based non-stationarity correction are being performed, as there may be a risk of such 'normalisation' behaving problematically for disconnected fixels.