fixelfilter: New filter connect
This performs a connected-components filter in a similar vein to maskfilter connect, in that in the output image, each fixel will be assigned an integer value that corresponds to a cluster index, with the largest clusters having the smallest indices.
Note that mechanistically, there are two separate thresholds applied within this connected-component filter: One for determining whether a fixel is super-threshold and hence should be included in a cluster, and another for testing the fractional connectivity between fixels to decide whether or not they should belong to the same cluster.