run.command(): Don't suppress MRtrix3 command outputs
Previously the -quiet flag was added to all MRtrix3 command calls, as shell=True was used in subprocess. Now that this is no longer the case, it is preferable to instead have all commands generate their standard amount of terminal output, so that script authors can parse this text if desired.
The additional -info flag passed when a script is run in debug mode will additionally check to make sure that no such flag is included in the command string itself, which could otherwise lead to a command-line parsing error in the underlying command.