mesh2voxel: Deal with ambiguous cases
During voxel-based segmentation of a surface-based structure, if there is sufficient ambiguity about whether or not a connected-componet cluster is inside or outside of the structure, try looking at the corners of the FoV to see if they are a part of the component: if all eight corners are either in or not in the connected component, use that to classify the connected component as being insode or outside the structure respectively. If there is still ambiguity, throw an exception.
This in particular fixes the mesh2voxel operation when a surface is thinner than the voxel size, which breaks certain assumptions that are used when classifying voxels nearby the surface.