Revert "5ttgen hsvs: Reorientation of T1 image"
This reverts commit 5e74d6ab539c49d2f9863fd0709859c42aab3f0c.

It was observed that in the "5ttgen_new_algos" branch, the -strides option in mrconvert was not working. By merging this branch with MRtrix3 tag 3.0_RC3, this problem appears to have been fixed. By doing so, the change introduced in the aforementioned commit - where the FSL command fslswapdim was explicitly used to force an image into LAS convention - should no longer be necessary.