Revert "5ttgen hsvs: Reorientation of T1 image"
This reverts commit 7a36ce87f28abe79a79ed85bac4eebb8c4b09374.

It was observed that in the "5ttgen_new_algos" branch, the -strides option in mrconvert was not working. By merging this branch with MRtrix3 tag 3.0_RC3, this problem appears to have been fixed. By doing so, the change introduced in the aforementioned commit - where the FSL command fslswapdim was explicitly used to force an image into LAS convention - should no longer be necessary.