Initial commit adding testing for Python scripts
A second submodule is added, which contains input whole-brain data for executing Python scripts, and output data as generated by the current master / dev branches.
Providing the -scripts option to the run_tests bash script will execute these tests instead of the CI tests of MRtrix3 binaries.
Many of these scripts depend on other neuroimaging software packages, and/or take a long time to execute, and therefore cannot be tested as part of continuous integration.