dwipreproc: Bug fix
dwipreproc performs explicit removal of unwanted header fields from the output DWI, as introduced in 47e7288f / #1389, by exporting the DWI header key-value pairs to a JSON file and explicitly editing the JSON contents. The writing of the input JSON file was however lost during the process of resolving divergent code in 76177871. This commit fixes the issue.
Correct operation of code for selecting & executing different versions of eddy - including if both CUDA and OpenMP versions are installed but the former does not run correctly - has been verified for the changes in #1449.