Fixel::Filter::Smooth: Reset images index position
Because this filter loops over the second image axis (in case there are multiple fixel quantitative values stored in the one image), this would result in the index position along that axis to be outside the domain of the fixel data file upon filter completion. This change simply resets the position back to zero, so that any calling code does not have to itself reset the position along this axis (particularly if the fixel data files are inherently 1D, in which case that code wouuld not make a lot of sense).