connectome2tck: Fix assertion failure / segfault
Memory for node volumes and centres of mass when generating exemplars was initialised based on the maximal node index observed in the assignments file. In the rare circumstance where one or more of the highest-valued nodes in a parcellation did not have any streamlines assigned to it, this would lead to a buffer over-run when compiled in release mode. This change introduces an explicit check for such, and issues a warning to the user.
Closes #1581.