mrtrix3.app: Change behaviour of trace() and var()
Alter these functions so that they behave identically to the TRACE() and VAR() functions used in the C++ library, which print to stderr regardless of verbosity level, and are used for active debugging. Usages of these functions in Python scripts have instead been replaced with calls to app.debug(), which is effectively how these functions were being used previously.