New Thread::run_ordered_queue() call
This add two functions to run a 3 & 4 stage pipeline respectively,
guaranteeing in-order processing of the items in the sink relative to
the source (though not in the pipe stages). This is done by buffering
items received out-of-order in the sink thread, until the missing items
are received, following by processing of buffered items. These new
functions impose certain conditions:

- source & sink must be single-threaded (otherwise impossible to impose
  ordering)
- if batching items, all items must be batched across all stages of the
  pipeline, with the same size batches
- if not batching items, all items must be passed through. Returning
  false from the pipe functor will abort the pipeline (batched operation
  allows skipping items in the pipe functor, since indexing is maintained
  per-batch).

Also further cleanup of existing Thread::run_queue() implementation.