run.command(): Don't validate input type
Because of the difficulties associated with types in which string data can be stored between Python2 and Python3, it is easiest to simply assume that if the input to run.command() is not a list, it is a type for which string-based operations can be performed.
Issue was observed running dwishellmath on Python2, as it constructs run.command() calls based on the output of image.mrinfo(), which comes out as unicode (yet my Python2 doesn't seem to recognise 'unicode' as a builtin keyword...).