Stats: Implement exchangeability blocks
As per the description in Winkler et al., 2014.
Inputs can be identified as having exchangeability that is either restricted to occur -within- individual blocks of observations, or for blocks containing equal numbers of observations where the -whole- blocks are exchanged.
In the case of whole-block exchangeability, sign-flipping of independent and symmetric errors can also be restricted to apply to whole blocks rather than individual observations.