permute_axes adapter: fix handling of additional axes
This allows the correct interpretation of negative axis indices in the
specification of the new axes, which is required to insert additional
axes (e.g. going from dimensions [ X Y Z 3 ] to [ X Y Z 1 3 ]). This
behaviour is required and documented for mrconvert's -axes option,
which currently doesn't work as stated (fixed with this change).