New class MR::Math::ZStatistic
The purpose of this class is to efficiently convert t-statistics and F-statistics to z-statistics. One or more lookup tables are generated, based on the degrees of freedom of the relevant distribution, where z-statistics corresponding to input t-statistics or F-statistcs are pre-calculated based on closed-form solutions; hermite interpolation is then applied on the values of that table.
This also includes function MR::Math::betaincreg(), which provides the _regularised_ incomplete Beta function.