Statistical inference: Use Z-transformed statistics
Throughout all statistical inference commands, output both the calculated statistic for each element, as well as the Z-transformation of such, based on the hypothesis and degrees of freedom (the latter of which may vary between different elements of the image). Furthermore, it is the Z-statistic that is provided to the relevant statistical enhancement algorithm, ensuring that such algorithms should behave comparably across F-tests and t-tests, and across experiments with different numbers of inputs.