Math::betaincreg(): Provide alternative implementation
This is a copy of the implementation found at: https://github.com/codeplea/incbeta.
This function can be used when Eigen 3.2 is installed, and therefore the unsupported function betainc() is not provided.
The configure script will detect the availability or absence of that function, and engage this alternative, datum-wise function implementation in the latter case.