FIX: avoid numpy dependency in connectomeedit
This commit changes the implementation of the internal functions of the
connectomeedit command in such a way that the numpy package is no longer
necessary.

Morover, this commit changes the functions used to load and save the
connectomes, which now are managed with the convenience functions from
the `mrtrix3.matrix` module.