Docs: Install: Add libpng
Support for PNG images was added in #1541, and support for that particular image handler will only be added during compilation if the relevant environment variable is set by the configure script. However instructions to install the relevant library package in order to enable this functionality had not yet been added.