dwipreproc: Permit rigid-body between DWI and SE-EPI
If the -align_seepi option is used, and the only difference between the DWI series and the SE-EPI series is a rigid-body transformation, then there is no need for the SE-EPI volumes to be explicitly re-gridded to match the DWI voxel grid; it can instead be safely concatenated with the first DWI b=0 image despite the difference in header transform. The header transform of the SE-EPI series is therefore replaced with that of the DWIs explcitly, but only in this specific instance (i.e. -align_seepi is used, and the only difference between the images is within the header transform).