app.py: Fix for Python 2/3 pylint
pylint 2.4.0 in Python3 gives a new warning "consider-using-sys-exit". By explicitly disabling this warning, pylint on Python 2 instead complains that this is a bad option value. Furthermore, the preferred exit() call for the signal handler invokes a protected function of the os module. This change removes the sys.exit() disable, and instead disables the protected-member warning.