tcken: Revert default maximum angle calculation
Revert changes that were made in #1507 with respect to determination of the default maximum angle for the various tracking algorithms. There, this decision was changed from being based on (90deg * step_size / vox()) to being a minimum radius of curvature of one voxel. Additionally, the default step size in the presence of RK4 was changed from 0.1*vox() to 0.5*vox(). This commit reverts these changes, but retains the calculation of minimum radius of curvature necessary for streamline length calculation optimisations.