dwipreproc: Fix -eddyqc_all for FSL6
in the updated FSL, eddy commands no longer produce output images relating to CNR maps and GP residuals automatically, instead requiring an explicit command-line option for each of these. This means that dwipreproc's -eddyqc_all option no longer writes these images to the output directory.
In this change, the eddy help pages are queried, and the relevant command-line options are provided to eddy only if necessary to produce the desired output.
In addition, mrtrix3.fsl.findImage() is not used to obtain these images in case of file suffix changes, as it throws an error if the image is not found; an explicit search for .nii and .nii.gz extensions are used.