connectome2tck: Skip non-valid exemplars
If an exemplar cannot be generated due to a parcel not being present in the index image, it is necessary to write an empty exemplar streamline; otherwise, propagation of the NaN values from the node centre of mass will be incorrectly interpreted as delimiters between streamlines in the track file.
Fixes #1769.