Changes to script testing
When testing scripts, particularly those reliant on external software packages, it is not always possible to perform a precise image difference comparison between output and pre-generated data. In such cases it may be preferable to retain the outputs of test commands using sensible naming in order to facilitate manual comparison. Therefore, instead of generating "tmp*" files within the data sub-directory and then deleting / over-writing them, create a new directory testing/scripts/tmp/, where the outputs of tested scripts are written. The structure is the same as the contents of testing/scripts/data/, such that cross-referencing should not be difficult. This change also carries the benefit of clarifying the operation of a number of more complex tests.