Final cleanup at end of run_tests
When performing multiple tests rather than just one, delete any residual temporary files after completion of the last test. This may prevent command execution errors due to library / executable mismatches due to untracked files in the submodule leading to a "dirty" version string suffix, and there is a chance that this may improve the tracking of submodule commit hashes when branch-hopping on the main repository. Note however that if only a solitary test is performed, whatever files it generates will not be removed.