New to<std::complex>() specialisation
A new algorithm for converting string data to complex numbers was necessary in order to make parsing of numerical input values in mrcalc usage more robust.
mrcalc will also now additionally give extra information whenever it is unable to interpret an input argument.
Explicit code in mrcalc for reading "nan", "-nan", "inf" and "-inf" was removed: the positive variants are handled appropriately using to<>(), whereas the latter are erroneously interpreted as command-line options prior to execution reaching this point.
A unit test for ensuring that the to<>() function does as expected in a wide range of scenarios was added.