run_tests: Change output filepaths
Because the "run_tests" script is now run multiple times on both CI platforms, the output logfile of the first execution (unit tests) would be moved to a new filesystem path before performing the second execution (binary tests). However in the case where at least one test fails, the bash "&&" operator does not execute the file move operation, such that the logfile is not appropriately stored as an artifact / concatenated for investigated.
This change makes the "run_tests" script change the path of the output log text file depending on the test being run, so that "./run_tests units" and "./run_tests binaries" intrinsically write to different file paths.