tckmap: Fix -ends_only for single-vertex streamlines
Function TrackMapperBase::voxelise_ends() intentionally allows a streamline for which both endpoints reside in the same voxel to increment the intensity of the output image by 2. However for single-vertex streamlines, the same vertex was being used as both the start and end point. This change ensures that for single-vertex streamlines, that one vertex contributes only once to the output image.
Fixes #1820.