dirflip and dirsplit: Fix command-line option parsing
Due to the use of scientific notation in the definitions of default values for the number of permutations in these two commands, the floating-point version of template function get_option_value<>() was silently invoked, rather than the intended integer version. This meant that any value passed by the user at the command-line would be interpreted as a floating-point number. In cases where a decimal multiplier was specified (e.g. 1B = one billion), this would not be interpreted correctly, as such relies on parsing as an integer. Previously, this usage would silently result in a value of 1, since the trailing "B" would be silently ignored. The issue was detected due to #1794: such usage now throws an error since conversion of string "1B" to floating-point fails to consume the full string. This commit fixes the parsing of specifically these two command-line options.