mrtrix_cleanup: Pylint / recursion fix
Fix syntax of code that prevents script from searching recursively within directories already identified for deletion. Not only did pylint flag an unnecessary comprehension, the recursion prevention was not working correctly, as the filter object would be fully consumed by the time a second attempt was made to read from it.