More robust read of FreeSurfer surface data
In the FreeSurfer surface format, the text header historically contains two lines, and hence parsers would look for these. However it appears that some files may contain only one newline character in the header, and these are still apparently valid (including some of the example data provided with FreeSufer). This change makes the parsing of this file format more robust to the different possibilities for the input data with respect to the contents of this header.