dwi2response: Fix handling of mrconvert_keyval argument
The problem is that path.from_user() quotes the path by default, which
is the right thing to do most of the time since its output is supposed
to be inserted into the command _string_. But when its output is used
as the mrconvert_keyval argument, it gets inserted into the command
_list_ (run.py: 315), and so shouldn't be quoted. Note this is only a
problem for paths that do need to be quoted, which is why this issue
hasn't come up so far.

More generally, there is a problem that:

     quote (quote ('some problematic string'))

will not match:

     quote ('some problematic string')

since it'll try to quote whatever additions the previous invocation did
to properly quote the original string. I'm not sure whether there's any
way around that...

Note that this will likely need replication in other commands too.