dwi2response tournier: -iter_voxels defaults to 10x -number
This changes means that if -number is set to a non-default value, the
number of voxels considered at the next iteration scales with it. It was
previously hard-wired to 3000, which clearly wouldn't be appropriate if
-number was 3000 or larger.

This also means that we can remove the explicit -iter_voxels option in
the dwi2reponse dhollander call when invoking an alternative -wm_algo,
without any change in its behaviour. This makes it possible to use the
fa algorithm (would previously fail due to unknown -iter_voxels option).

Note that it still can't use the tax algorithm, since it can't handle
the -number option.