5ttgen hsvs: Improve mapping of brainstem to volume 4
Immediate addition of the brain stem segmentation to volume index 4 resulted in other segmented structures in the vicinity of the brain stem being removed, as the tissue combination logic prioritises the segmentation of pathological structures. This change instead initially maps the brain stem to the white matter, such that other structures may receive priority in the segmentation, and then moves any remaining white matter in that area to volume 4 as a final step.