5ttgen hsvs: Improve lateral ventricles
Rather than dealing with lateral ventricles and choroid plexuses independently, with conversion to mesh / smoothing / mapping to partial volume fractions, instead merge the two masks together prior to mesh conversion. This prevents the appearance of small pieces of white matter in the middle of the ventricles where the two resulting surfaces did not quite intersect.