Header::create(): Don't read 4D schemes from cmdline options
Following 52175b3a (#1555), on running dwiextract using the -fslgrad option, any output image would not contain a gradient table for the extracted data. This was because Header::create() now calls DWI::get_DW_scheme() and wipes the gradient table on a raised Exception; in turn, that function was raising an Exception in DWI::load_bvecs_bvals() as the file loaded via -fslgrad did not match the output image, for which a reduced gradient table had already been constructed and stored in the header. To correct this, Header::create() will now call DWI::parse_DW_scheme() rather than DWI::get_DW_scheme(), so it never attempts to reader from the -grad or -fslgrad options. Equivalent changes are also made for phase encoding information.