dwifslpreproc: Fix writing of matrix files for FSL
For #1603, the intention was to not write any comments at the header of files that were being generated for the purpose of feeding to FSL commands. However the relevant function calls were not updated to reflect the final implementation of the matrix.save_*() functions. This fix simply disables the addition of the command history string to the header of the file, the result of which is that no header information is written.