matrix.save_numeric(): Fix for erroneous usage
If matrix.save_numeric() is run whilst providing either the "header" or "footer" keyword arguments, and the contents of that argument was either an empty string or an empty list, this would lead to cryptic errors later in the function due to attempting to perform a dictionary function on a non-dictionary variable.
Erroneously omitted from #1886.