Remove all references to "output_step_size"
This field was intended to inform decisions regarding length calculations, but is not adequately robust to serve that purpose. For instance, use of upsampling / downsampling results in a non-constant step size. Until some more robust technique is used, this parameter perhaps gives a false sense of certainty. It does however mean that if the distance between vertices in the track file differs substantially from the tracking step size, some calculations may be inaccurate; though those are typically not calculations where precision is essential.
CLoses #1813.