fsl.exe_name(): Warn if using "fsl5.0-" prefix
Follows from #1459.
While in some cases, use of the "fsl5.0-" prefixed commands is apparently a convenience, in others it has been demonstrated that use of such commands can lead to either undesirable behaviour, or use of an unexpected software version. While use of these binaries has been relegated in the priority order, it may nevertheless be indicative of a problem if such are used. Therefore, this change introduces a warning to users if such executables are utilised, but will attempt to continue processing. Even if use of these binaries is not problematic, it nevertheless warrants attention from the user.