fixel2voxel none: Revert default fill to 0.0
Placing values of NaN in voxels where there are fewer fixels than the maximum resulted in altered behaviour of the dwi2response tournier algorithm in specific instances. This returns the default behaviour of this operation to be the same as the previous "fixel2voxel split_data" (changes in #1874). There is already an option "-fill" in place to change this at the command-line; only the default behaviour is being changed.