Stats: Modify import of cohort data
Loading statistical inference cohort data could lead to issues if the text file containing the subject file list resided in a different directory. fixelcfestats was unique in that it would also look for those files relative to the input fixel directory; but this would not work if the subject list text file was not in the current working directory, as the directory of that file was erroneously added to the file basename prior to Path::join(). In this change, the potential utilisation of different base directories in order to find the input data is generalised and moved to CohortDataImport::initialise(). For most commands, locating all input files is first tested relative to the location of the subjects test file, then the current working directory. In the case of fixelcfestats, the input fixel directory tested before either of these two. Only if all input files are found on the filesystem for a given directory location is the import of data from those files attempted.